remove happy path tests and move contents to simple integration tests for compile and test
