+++ packages/truffle/test/scenarios/commands/test.js
@@ -1,0 +1,28 @@
+const sandbox = require("../sandbox");
+const MemoryLogger = require("../MemoryLogger");
+const CommandRunner = require("../commandRunner");
+const path = require("path");
+const { assert } = require("chai");
+const Server = require("../server");
+
+describe("truffle test", function () {
+  let config;
+  const logger = new MemoryLogger();
+  const project = path.join(__dirname, "../../sources/toyProject");
+
+  before(async () => {
+    await Server.start();
+    config = await sandbox.create(project);
+    config.logger = logger;
+  });
+  after(async function () {
+    await Server.stop();
+  });
+
+  it("runs tests", async function () {
+    this.timeout(70000);
+    await CommandRunner.run("test", config);
+    const output = logger.contents();
+    assert(output.indexOf("0 passing") >= 0);
+  });
+});
